vllm/pyproject.toml
[tool.poetry]
name = "hades-data-llm"
version = "0.1.0"
description = "Harnessing AI to Disrupt and Evaluate Security (HADES)"
authors = ["cyberphor <victor.fernandeziii@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
autogen-agentchat = ">=0.2,<0.4"
fastapi = "^0.115.6"
llama-cpp-python = "^0.3.5"
huggingface-hub = "^0.27.0"
numpy = ">=1.26"
uvicorn = "^0.34.0"
websockets = "^14.1"

[tool.poetry.scripts]
hades = "hades.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
